+++ src/Resource/AccountBilling.php
@@ -66,5 +66,5 @@ class AccountBilling extends BaseResource
 
     public bool $automaticTagsPlacement;
 
-    public bool $multiSign;
+    public bool $bulkSigning;
 }