+++ tests/DigiSignTest.php
@@ -156,8 +156,8 @@ class DigiSignTest extends TestCase
         self::assertSame('/api/my', $mockClient->getLastRequest()->getUri()->getPath());
         $dgs->report()->request('GET');
         self::assertSame('/api/report', $mockClient->getLastRequest()->getUri()->getPath());
-        $dgs->multiSign()->request('GET');
-        self::assertSame('/api/multi-signs', $mockClient->getLastRequest()->getUri()->getPath());
+        $dgs->bulkSignature()->request('GET');
+        self::assertSame('/api/bulk-signatures', $mockClient->getLastRequest()->getUri()->getPath());
     }
 
     public function testUserAgent(): void